Refactored vote results

+++ Blog.Business/Models/QuestionaryAnswerModels/Statistics.cs
@@ -4,6 +4,6 @@
     {
         public string Answer { get; set; }
         public int AnsweredCount { get; set; }
-        public int AnsweredPercent { get; set; }
+        public double AnsweredPercent { get; set; }
     }
 }
